First version including meta alg approach. Still not fully implemented, but the system is ready.
The idea is that a task can contain other tasks as jobs. So the
MetaAligner tasks includes normal jobs (reversing algs) and, for
instance, Muscle and Mafft jobs.

Proper bindings to all single aligners are to be written. Also,
general design of nested Tasks and Jobs needs a lot cleaning.