First real config with meta-alg and 2 aligners. Fixed nested tasks.
Nested tasks were failing when subtaks need to call finish() function
to complete the job, so I added a proper exec_jobs() function in
master_task. This solves the problem of inplace execution, but I will
possible need to change it to support external job
submission. Subtasks are not scheduled, so, so far, they can only run
inplace.