Fixed problem of dependent jobs. Curses interface is no longer the default mode (use -U to enable). Meta_aligner is now fully working.
Task jobs can now depend from each other. The system is clever enough
to submit first the non dependent jobs, when they are finished the
next round, and so on.

This allows to execute inplace or to submit rounds of commands to
other machines. When they are done, next round can be submitted.