Added tree annotation, Uhire task. Changed taskdir behavior:
Added a rudimentary tree annotation method, executed every time a task
is finished.

Added Uhire task to deal with huge alignments.

Changed the way I create taskdirs. Before, taskdir depended on
child-job parameters. Now taskdir is a random string and a __id__ file
is created. In the future, I will have to implement a system to check
if a task is already executed.